+++ sydtest/src/Test/Syd/Modify.hs
@@ -23,10 +23,17 @@ module Test.Syd.Modify
     withExecutionOrderRandomisation,
     ExecutionOrderRandomisation (..),
 
+    -- * Modifying the number of retries
+    modifyRetries,
+    withoutRetries,
+    withRetries,
+
     -- * Declaring flakiness
     flaky,
     flakyWith,
     notFlaky,
+    potentiallyFlaky,
+    potentiallyFlakyWith,
     withFlakiness,
     FlakinessMode (..),
   )
@@ -77,6 +84,18 @@ randomiseExecutionOrder = withExecutionOrderRandomisation RandomiseExecutionOrde
 withExecutionOrderRandomisation :: ExecutionOrderRandomisation -> TestDefM a b c -> TestDefM a b c
 withExecutionOrderRandomisation p = censor ((: []) . DefRandomisationNode p)
 
+-- | Modify the number of retries to use in flakiness diagnostics.
+modifyRetries :: (Word -> Word) -> TestDefM a b c -> TestDefM a b c
+modifyRetries modRetries = censor ((: []) . DefRetriesNode modRetries)
+
+-- | Turn off retries
+withoutRetries :: TestDefM a b c -> TestDefM a b c
+withoutRetries = modifyRetries (const 0)
+
+-- | Make the number of retries this constant
+withRetries :: Word -> TestDefM a b c -> TestDefM a b c
+withRetries w = modifyRetries (const w)
+
 -- | Mark a test suite as "potentially flaky".
 --
 -- This will retry any test in the given test group up to the given number of tries, and pass a test if it passes once.
@@ -86,22 +105,31 @@ withExecutionOrderRandomisation p = censor ((: []) . DefRandomisationNode p)
 -- In other words: tests using flaky must be guaranteed to fail every time if
 -- an error is introduced in the code, it should only be added to deal with
 -- accidental failures, never accidental passes.
-flaky :: Int -> TestDefM a b c -> TestDefM a b c
-flaky i = withFlakiness $ MayBeFlakyUpTo i Nothing
+flaky :: Word -> TestDefM a b c -> TestDefM a b c
+flaky i = modifyRetries (const i) . withFlakiness (MayBeFlaky Nothing)
 
 -- | Like 'flaky', but also shows the given message to the user whenever the test is flaky.
 --
 -- You could use it like this:
 --
 -- >>> flakyWith 3 "Something sometimes goes wrong with the database, see issue 6346" ourTestSuite
-flakyWith :: Int -> String -> TestDefM a b c -> TestDefM a b c
-flakyWith i message = withFlakiness $ MayBeFlakyUpTo i (Just message)
+flakyWith :: Word -> String -> TestDefM a b c -> TestDefM a b c
+flakyWith i message = modifyRetries (const i) . withFlakiness (MayBeFlaky (Just message))
 
 -- | Mark a test suite as "must not be flaky".
 --
 -- This is useful to have a subgroup of a group marked as 'flaky' that must not be flaky afteral.
 notFlaky :: TestDefM a b c -> TestDefM a b c
 notFlaky = withFlakiness MayNotBeFlaky
+
+-- | Mark a test suite as 'potentially flaky', such that it will not fail if it is
+-- flaky but passes at least once.
+potentiallyFlaky :: TestDefM a b c -> TestDefM a b c
+potentiallyFlaky = withFlakiness (MayBeFlaky Nothing)
+
+-- | Like 'potentiallyFlaky', but with a message.
+potentiallyFlakyWith :: String -> TestDefM a b c -> TestDefM a b c
+potentiallyFlakyWith message = withFlakiness (MayBeFlaky (Just message))
 
 -- | Annotate a test group with 'FlakinessMode'.
 withFlakiness :: FlakinessMode -> TestDefM a b c -> TestDefM a b c